Move Debian and Ubuntu to DistributionInstaller
